move edit link to the index
prevents it from showing up in the "Authorizing" modal
